+++ integrationtest/backend/tck/src/main/java/org/hibernate/search/integrationtest/backend/tck/search/query/SearchQueryFetchIT.java
@@ -84,7 +84,7 @@ public class SearchQueryFetchIT {
 
 	@Test
 	public void fetch_limit() {
-		assertThat( matchAllQuery().fetch( (Integer) null ) )
+		assertThat( matchAllQuery().fetch(null) )
 				.hasTotalHitCount( DOCUMENT_COUNT )
 				.hasDocRefHitsExactOrder( builder -> {
 					for ( int i = 0; i < DOCUMENT_COUNT; i++ ) {
@@ -152,7 +152,7 @@ public class SearchQueryFetchIT {
 
 	@Test
 	public void fetchHits_limit() {
-		assertThat( matchAllQuery().fetchHits( (Integer) null ) )
+		assertThat( matchAllQuery().fetchHits(null) )
 				.hasDocRefHitsExactOrder( builder -> {
 					for ( int i = 0; i < DOCUMENT_COUNT; i++ ) {
 						builder.doc( INDEX_NAME, docId( i ) );
